Task-07: add aesthetic loader
