+++ src/components/loader/loader.module.css
@@ -1,0 +1,110 @@
+.loaderWrap {
+  width: 100%;
+  height: 80vh;
+  display: flex;
+  justify-content: center;
+  align-items: center;
+}
+
+.ldSpinner,
+.ldSpinner div,
+.ldSpinner div::after {
+  box-sizing: border-box;
+}
+
+.ldSpinner {
+  color: var(--orange-color);
+  display: inline-block;
+  position: relative;
+  width: 80px;
+  height: 80px;
+}
+
+.ldSpinner div {
+  transform-origin: 40px 40px;
+  animation: ldSpinner 1.2s linear infinite;
+}
+
+.ldSpinner div::after {
+  content: "";
+  display: block;
+  position: absolute;
+  top: 3.2px;
+  left: 36.8px;
+  width: 6.4px;
+  height: 17.6px;
+  border-radius: 20%;
+  background: currentcolor;
+}
+
+.ldSpinner div:nth-child(1) {
+  transform: rotate(0deg);
+  animation-delay: -1.1s;
+}
+
+.ldSpinner div:nth-child(2) {
+  transform: rotate(30deg);
+  animation-delay: -1s;
+}
+
+.ldSpinner div:nth-child(3) {
+  transform: rotate(60deg);
+  animation-delay: -.9s;
+}
+
+.ldSpinner div:nth-child(4) {
+  transform: rotate(90deg);
+  animation-delay: -.8s;
+}
+
+.ldSpinner div:nth-child(5) {
+  transform: rotate(120deg);
+  animation-delay: -.7s;
+}
+
+.ldSpinner div:nth-child(6) {
+  transform: rotate(150deg);
+  animation-delay: -.6s;
+}
+
+.ldSpinner div:nth-child(7) {
+  transform: rotate(180deg);
+  animation-delay: -.5s;
+}
+
+.ldSpinner div:nth-child(8) {
+  transform: rotate(210deg);
+  animation-delay: -.4s;
+}
+
+.ldSpinner div:nth-child(9) {
+  transform: rotate(240deg);
+  animation-delay: -.3s;
+}
+
+.ldSpinner div:nth-child(10) {
+  transform: rotate(270deg);
+  animation-delay: -.2s;
+}
+
+.ldSpinner div:nth-child(11) {
+  transform: rotate(300deg);
+  animation-delay: -.1s;
+}
+
+.ldSpinner div:nth-child(12) {
+  transform: rotate(330deg);
+  animation-delay: 0s;
+}
+
+@keyframes ldSpinner {
+
+  0% {
+    opacity: 1;
+  }
+
+  100% {
+    opacity: 0;
+  }
+}
+
