+++ src/main.tsx
@@ -1,15 +1,15 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { HashRouter } from 'react-router-dom';
-import { ToastContainer } from 'react-toastify';
 
 import App from './App.tsx';
 
 import './index.css';
 
 ReactDOM.createRoot(document.querySelector('#root')!).render(
-    <HashRouter>
-        <App />
-        <ToastContainer />
-    </HashRouter>,
+    <React.StrictMode>
+        <HashRouter>
+            <App />
+        </HashRouter>
+    </React.StrictMode>,
 );
